Allow skip of ambient_removal by introducing method 'none'
The param.ambient_removal is used within work flow AMBIENT_RNA_REMOVAL
(see subworkflows/local/ambient_rna_removal.nf) to skip the invocation
of that external tool.

This is meant as a quick fix when reads need to be filtered to perform that
removal since these removal methods are sensitive to cells with low quality data,
but only unfiltered data are available.